+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [] -
 ### Added
 ### Changed
+- Save Ensembl ids for demo patient data genes automatically
+- Mock the Ensembl REST API converting gene symbols to Ensembl IDs in tests
 ### Fixed
 - Syntax in the disclaimer